Prefer a software decoder before ExoPlayer for odd-size clips.
Video.Guru 48x570 reports format_supported=YES then dies in the hardware
MediaCodec renderer. Hook MediaCodecSelector.DEFAULT so software codecs
are first when opening those files, and retry the same clip instead of
the crash sheet if hardware still fails.

+++ lib/main.dart
@@ -91,7 +91,7 @@ class _VideoPlayerAppState extends State<VideoPlayerApp> with WidgetsBindingObse
             ThemeModePref.dark => ThemeMode.dark,
           },
           builder: (context, child) {
-            final scale = appSettings.uiScale.clamp(0.85, 1.35);
+            final scale = appSettings.uiScale.clamp(0.85, 1.35).toDouble();
             return MediaQuery(
               data: MediaQuery.of(context).copyWith(
                 textScaler: TextScaler.linear(scale),
